String splitter refactoring (#56889)
