+++ src/EditorFeatures/CSharp/SplitStringLiteral/SplitStringLiteralCommandHandler.cs
@@ -8,6 +8,7 @@ using System;
 using System.ComponentModel.Composition;
 using System.Linq;
 using System.Threading;
+using Microsoft.CodeAnalysis.CSharp.SplitStringLiteral;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Formatting;
@@ -19,6 +20,7 @@ using Microsoft.VisualStudio.Language.Intellisense.AsyncCompletion;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Editor.Commanding.Commands;
+using Microsoft.VisualStudio.Text.Editor.OptionsExtensionMethods;
 using Microsoft.VisualStudio.Text.Operations;
 using Microsoft.VisualStudio.Utilities;
 using Roslyn.Utilities;
@@ -32,15 +34,18 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
     internal partial class SplitStringLiteralCommandHandler : ICommandHandler<ReturnKeyCommandArgs>
     {
         private readonly ITextUndoHistoryRegistry _undoHistoryRegistry;
+        private readonly IGlobalOptionService _globalOptions;
         private readonly IEditorOperationsFactoryService _editorOperationsFactoryService;
 
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
         public SplitStringLiteralCommandHandler(
             ITextUndoHistoryRegistry undoHistoryRegistry,
+            IGlobalOptionService globalOptions,
             IEditorOperationsFactoryService editorOperationsFactoryService)
         {
             _undoHistoryRegistry = undoHistoryRegistry;
+            _globalOptions = globalOptions;
             _editorOperationsFactoryService = editorOperationsFactoryService;
         }
 
@@ -54,6 +59,11 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
 
         public bool ExecuteCommandWorker(ReturnKeyCommandArgs args)
         {
+            if (!_globalOptions.GetOption(SplitStringLiteralOptions.Enabled, LanguageNames.CSharp))
+            {
+                return false;
+            }
+
             var textView = args.TextView;
             var subjectBuffer = args.SubjectBuffer;
             var spans = textView.Selection.GetSnapshotSpansOnBuffer(subjectBuffer);
@@ -83,13 +93,16 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
                 }
             }
 
+            var useTabs = !textView.Options.IsConvertTabsToSpacesEnabled();
+            var tabSize = textView.Options.GetTabSize();
+
             // We now go through the verified string literals and split each of them.
             // The list of spans is traversed in reverse order so we do not have to
             // deal with updating later caret positions to account for the added space
             // from splitting at earlier caret positions.
             foreach (var span in spans.Reverse())
             {
-                if (!SplitString(textView, subjectBuffer, span.Start))
+                if (!SplitString(textView, subjectBuffer, span.Start.Position, useTabs, tabSize, CancellationToken.None))
                 {
                     return false;
                 }
@@ -98,41 +111,44 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
             return true;
         }
 
-        private bool SplitString(ITextView textView, ITextBuffer subjectBuffer, SnapshotPoint caret)
+        private bool SplitString(ITextView textView, ITextBuffer subjectBuffer, int position, bool useTabs, int tabSize, CancellationToken cancellationToken)
         {
             var document = subjectBuffer.CurrentSnapshot.GetOpenDocumentInCurrentContextWithChanges();
-
-            if (document != null)
+            if (document == null)
             {
-                var options = document.GetOptionsAsync(CancellationToken.None).WaitAndGetResult(CancellationToken.None);
-                var enabled = options.GetOption(SplitStringLiteralOptions.Enabled);
-
-                if (enabled)
-                {
-                    using var transaction = CaretPreservingEditTransaction.TryCreate(
-                        CSharpEditorResources.Split_string, textView, _undoHistoryRegistry, _editorOperationsFactoryService);
-
-                    var cursorPosition = SplitStringLiteral(document, options, caret, CancellationToken.None);
-                    if (cursorPosition != null)
-                    {
-                        var snapshotPoint = new SnapshotPoint(
-                            subjectBuffer.CurrentSnapshot, cursorPosition.Value);
-                        var newCaretPoint = textView.BufferGraph.MapUpToBuffer(
-                            snapshotPoint, PointTrackingMode.Negative, PositionAffinity.Predecessor,
-                            textView.TextBuffer);
-
-                        if (newCaretPoint != null)
-                        {
-                            textView.Caret.MoveTo(newCaretPoint.Value);
-                        }
-
-                        transaction.Complete();
-                        return true;
-                    }
-                }
+                return false;
             }
 
-            return false;
+            // TODO: read option from textView.Options (https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1412138)
+            var indentStyle = document.Project.Solution.Options.GetOption(FormattingOptions.SmartIndent, LanguageNames.CSharp);
+
+            using var transaction = CaretPreservingEditTransaction.TryCreate(
+                CSharpEditorResources.Split_string, textView, _undoHistoryRegistry, _editorOperationsFactoryService);
+
+            var splitter = StringSplitter.TryCreate(document, position, useTabs, tabSize, indentStyle, cancellationToken);
+            if (splitter?.TrySplit(out var newDocument, out var newPosition) != true)
+            {
+                return false;
+            }
+
+            // apply the change:
+            var workspace = newDocument.Project.Solution.Workspace;
+            workspace.TryApplyChanges(newDocument.Project.Solution);
+
+            // move caret:
+            var snapshotPoint = new SnapshotPoint(
+                subjectBuffer.CurrentSnapshot, newPosition);
+            var newCaretPoint = textView.BufferGraph.MapUpToBuffer(
+                snapshotPoint, PointTrackingMode.Negative, PositionAffinity.Predecessor,
+                textView.TextBuffer);
+
+            if (newCaretPoint != null)
+            {
+                textView.Caret.MoveTo(newCaretPoint.Value);
+            }
+
+            transaction.Complete();
+            return true;
         }
 
         private static bool LineContainsQuote(ITextSnapshotLine line, int caretPosition)
@@ -147,27 +163,6 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
             }
 
             return false;
-        }
-
-        private static int? SplitStringLiteral(
-            Document document, DocumentOptionSet options, int position, CancellationToken cancellationToken)
-        {
-            var useTabs = options.GetOption(FormattingOptions.UseTabs);
-            var tabSize = options.GetOption(FormattingOptions.TabSize);
-            var indentStyle = options.GetOption(FormattingOptions.SmartIndent, LanguageNames.CSharp);
-
-            var root = document.GetSyntaxRootSynchronously(cancellationToken);
-            var sourceText = root.SyntaxTree.GetText(cancellationToken);
-
-            var splitter = StringSplitter.Create(
-                document, position, root, sourceText,
-                useTabs, tabSize, indentStyle, cancellationToken);
-            if (splitter == null)
-            {
-                return null;
-            }
-
-            return splitter.TrySplit();
         }
     }
 }
