+++ src/Features/CSharp/Portable/SplitStringLiteral/InterpolatedStringSplitter.cs
@@ -7,16 +7,15 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Text;
-using static Microsoft.CodeAnalysis.Formatting.FormattingOptions;
 
-namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
+namespace Microsoft.CodeAnalysis.CSharp.SplitStringLiteral
 {
-    internal partial class SplitStringLiteralCommandHandler
+    internal abstract partial class StringSplitter
     {
-        private class InterpolatedStringSplitter : StringSplitter
+        private sealed class InterpolatedStringSplitter : StringSplitter
         {
             private readonly InterpolatedStringExpressionSyntax _interpolatedStringExpression;
 
@@ -24,7 +23,7 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.SplitStringLiteral
                 Document document, int position,
                 SyntaxNode root, SourceText sourceText,
                 InterpolatedStringExpressionSyntax interpolatedStringExpression,
-                bool useTabs, int tabSize, IndentStyle indentStyle,
+                bool useTabs, int tabSize, FormattingOptions.IndentStyle indentStyle,
                 CancellationToken cancellationToken)
                 : base(document, position, root, sourceText, useTabs, tabSize, indentStyle, cancellationToken)
             {
